Refactoring and support for universal media types

+++ src/NzbDrone.Core/Download/Pending/PendingReleaseService.cs
@@ -32,38 +32,40 @@ namespace NzbDrone.Core.Download.Pending
 
     public class PendingReleaseService : IPendingReleaseService,
                                          IHandle<SeriesDeletedEvent>,
-                                         IHandle<EpisodeGrabbedEvent>,
-                                         IHandle<MovieGrabbedEvent>,
+                                         IHandle<RemoteItemGrabbedEvent>,
                                          IHandle<RssSyncCompleteEvent>
     {
         private readonly IIndexerStatusService _indexerStatusService;
         private readonly IPendingReleaseRepository _repository;
         private readonly ISeriesService _seriesService;
-        private readonly IParsingService _parsingService;
+        private readonly IParsingServiceProvider _parsingServiceProvider;
         private readonly IDelayProfileService _delayProfileService;
         private readonly ITaskManager _taskManager;
         private readonly IConfigService _configService;
         private readonly IEventAggregator _eventAggregator;
+        private readonly IMovieService _movieService;
         private readonly Logger _logger;
 
         public PendingReleaseService(IIndexerStatusService indexerStatusService,
                                     IPendingReleaseRepository repository,
                                     ISeriesService seriesService,
-                                    IParsingService parsingService,
+                                    IParsingServiceProvider parsingServiceProvider,
                                     IDelayProfileService delayProfileService,
                                     ITaskManager taskManager,
                                     IConfigService configService,
                                     IEventAggregator eventAggregator,
+                                    IMovieService movieService,
                                     Logger logger)
         {
             _indexerStatusService = indexerStatusService;
             _repository = repository;
             _seriesService = seriesService;
-            _parsingService = parsingService;
+            _parsingServiceProvider = parsingServiceProvider;
             _delayProfileService = delayProfileService;
             _taskManager = taskManager;
             _configService = configService;
             _eventAggregator = eventAggregator;
+            _movieService = movieService;
             _logger = logger;
         }
 
@@ -72,13 +74,13 @@ namespace NzbDrone.Core.Download.Pending
         {
             var alreadyPending = GetPendingReleases();
 
-            var episodeIds = decision.RemoteEpisode.Episodes.Select(e => e.Id);
+            var ids = decision.Item.GetItemIds();
 
-            var existingReports = alreadyPending.Where(r => r.RemoteEpisode.Episodes.Select(e => e.Id)
-                                                             .Intersect(episodeIds)
+            var existingReports = alreadyPending.Where(r => r.GetPendingReleasesIds()
+                                                             .Intersect(ids)
                                                              .Any());
 
-            if (existingReports.Any(MatchingReleasePredicate(decision.RemoteEpisode.Release)))
+            if (existingReports.Any(MatchingReleasePredicate(decision.Item.Release)))
             {
                 _logger.Debug("This release is already pending, not adding again");
                 return;
@@ -120,36 +122,53 @@ namespace NzbDrone.Core.Download.Pending
 
             foreach (var pendingRelease in GetPendingReleases())
             {
-                foreach (var episode in pendingRelease.RemoteEpisode.Episodes)
-                {
-                    var ect = pendingRelease.Release.PublishDate.AddMinutes(GetDelay(pendingRelease.RemoteEpisode));
-
-                    if (ect < nextRssSync.Value)
+                queued.AddRange(pendingRelease.RemoteItem.ForEachMediaItem(
+                    (remoteEpisode, episode) =>
                     {
-                        ect = nextRssSync.Value;
-                    }
-                    else
-                    {
-                        ect = ect.AddMinutes(_configService.RssSyncInterval);
-                    }
+                        var ect = remoteEpisode.Release.PublishDate.AddMinutes(GetDelay(remoteEpisode));
 
-                    var queue = new Queue.Queue
-                                {
-                                    Id = GetQueueId(pendingRelease, episode),
-                                    Series = pendingRelease.RemoteEpisode.Series,
-                                    Episode = episode,
-                                    Quality = pendingRelease.RemoteEpisode.ParsedEpisodeInfo.Quality,
-                                    Title = pendingRelease.Title,
-                                    Size = pendingRelease.RemoteEpisode.Release.Size,
-                                    Sizeleft = pendingRelease.RemoteEpisode.Release.Size,
-                                    RemoteEpisode = pendingRelease.RemoteEpisode,
-                                    Timeleft = ect.Subtract(DateTime.UtcNow),
-                                    EstimatedCompletionTime = ect,
-                                    Status = "Pending",
-                                    Protocol = pendingRelease.RemoteEpisode.Release.DownloadProtocol
-                                };
-                    queued.Add(queue);
-                }
+                        ect = ect < nextRssSync.Value ? nextRssSync.Value : ect.AddMinutes(_configService.RssSyncInterval);
+
+                        return new Queue.Queue
+                        {
+                            Id = episode.GetQueueId(pendingRelease),
+                            Series = remoteEpisode.Series,
+                            Episode = episode,
+                            Quality = remoteEpisode.Info.Quality,
+                            Title = pendingRelease.Title,
+                            Size = remoteEpisode.Release.Size,
+                            Sizeleft = remoteEpisode.Release.Size,
+                            RemoteItem = remoteEpisode,
+                            Timeleft = ect.Subtract(DateTime.UtcNow),
+                            EstimatedCompletionTime = ect,
+                            Status = "Pending",
+                            Protocol = remoteEpisode.Release.DownloadProtocol
+                        };
+                    },
+                    remoteMovie =>
+                    {
+                        var ect = remoteMovie.Release.PublishDate.AddMinutes(GetDelay(remoteMovie));
+
+                        ect = ect < nextRssSync.Value ? nextRssSync.Value : ect.AddMinutes(_configService.RssSyncInterval);
+
+                        return new Queue.Queue
+                        {
+                            Id = remoteMovie.Movie.GetQueueId(pendingRelease),
+                            Series = null,
+                            Episode = null,
+                            Movie = remoteMovie.Movie,
+                            Quality = remoteMovie.Info.Quality,
+                            Title = pendingRelease.Title,
+                            Size = remoteMovie.Release.Size,
+                            Sizeleft = remoteMovie.Release.Size,
+                            RemoteItem = remoteMovie,
+                            Timeleft = ect.Subtract(DateTime.UtcNow),
+                            EstimatedCompletionTime = ect,
+                            Status = "Pending",
+                            Protocol = remoteMovie.Release.DownloadProtocol
+                        };
+                    }
+                ));
             }
 
             //Return best quality release for each episode
@@ -173,13 +192,23 @@ namespace NzbDrone.Core.Download.Pending
         public void RemovePendingQueueItems(int queueId)
         {
             var targetItem = FindPendingRelease(queueId);
-            var seriesReleases = _repository.AllBySeriesId(targetItem.SeriesId);
 
-            var releasesToRemove = seriesReleases.Where(
-                c => c.ParsedEpisodeInfo.SeasonNumber == targetItem.ParsedEpisodeInfo.SeasonNumber &&
-                     c.ParsedEpisodeInfo.EpisodeNumbers.SequenceEqual(targetItem.ParsedEpisodeInfo.EpisodeNumbers));
+            var targetEpisode = targetItem as EpisodePendingRelease;
+            if (targetEpisode != null)
+            {
+                var seriesReleases = _repository.AllBySeriesId(targetItem.Dao.SeriesId);
 
-            _repository.DeleteMany(releasesToRemove.Select(c => c.Id));
+                var releasesToRemove = seriesReleases.Where(
+                    c => c.ParsedEpisodeInfo.SeasonNumber == targetItem.Dao.ParsedEpisodeInfo.SeasonNumber &&
+                         c.ParsedEpisodeInfo.EpisodeNumbers.SequenceEqual(
+                             targetItem.Dao.ParsedEpisodeInfo.EpisodeNumbers));
+
+                _repository.DeleteMany(releasesToRemove.Select(c => c.Id));
+            }
+            else
+            {
+                throw new NotImplementedException();
+            }
         }
 
         public RemoteEpisode OldestPendingRelease(int seriesId, IEnumerable<int> episodeIds)
@@ -189,38 +218,56 @@ namespace NzbDrone.Core.Download.Pending
                                                      .FirstOrDefault();
         }
 
-        private List<PendingRelease> GetPendingReleases()
+        private List<IPendingRelease> GetPendingReleases()
         {
-            var result = new List<PendingRelease>();
+            var result = new List<IPendingRelease>();
 
             foreach (var release in _repository.All())
             {
-                var remoteEpisode = GetRemoteEpisode(release);
+                // is it a movie or tv show release
+                var remoteItem = release.MovieId != 0 ? GetRemoteMovie(release) : GetRemoteEpisode(release);
+                if (remoteItem == null) continue;
 
-                if (remoteEpisode == null) continue;
-
-                release.RemoteEpisode = remoteEpisode;
-
-                result.Add(release);
+                var pendingRelease = release.MovieId != 0
+                    ? (IPendingRelease)new MoviePendingRelease(release, remoteItem)
+                    : (IPendingRelease)new EpisodePendingRelease(release, remoteItem);
+                result.Add(pendingRelease);
             }
 
             return result;
         }
 
+        private RemoteItem GetRemoteMovie(PendingRelease release)
+        {
+            var movie = _movieService.GetMovie(release.MovieId);
+
+            if (movie == null) return null;
+
+            return new RemoteMovie()
+            {
+                Release = release.Release,
+                Media = movie
+            };
+
+
+        }
+
         private RemoteEpisode GetRemoteEpisode(PendingRelease release)
         {
+            var parsingService = _parsingServiceProvider.GetTvShowParsingService();
             var series = _seriesService.GetSeries(release.SeriesId);
 
             //Just in case the series was removed, but wasn't cleaned up yet (housekeeper will clean it up)
-            if (series == null) return null;
+            if (series == null)
+                return null;
 
-            var episodes = _parsingService.GetEpisodes(release.ParsedEpisodeInfo, series, true);
+            var episodes = parsingService.GetEpisodes(release.ParsedEpisodeInfo, series, true);
 
             return new RemoteEpisode
             {
-                Series = series,
+                Media = series,
                 Episodes = episodes,
-                ParsedEpisodeInfo = release.ParsedEpisodeInfo,
+                Info = release.ParsedEpisodeInfo,
                 Release = release.Release
             };
         }
@@ -229,59 +276,63 @@ namespace NzbDrone.Core.Download.Pending
         {
             _repository.Insert(new PendingRelease
             {
-                SeriesId = decision.RemoteEpisode.Series.Id,
-                ParsedEpisodeInfo = decision.RemoteEpisode.ParsedEpisodeInfo,
-                Release = decision.RemoteEpisode.Release,
-                Title = decision.RemoteEpisode.Release.Title,
+                SeriesId = decision.Item.GetSeriesSafely()?.Id ?? 0,
+                MovieId = decision.Item.GetMovieSafely()?.Id ?? 0,
+                ParsedEpisodeInfo = (decision.Item.Info as ParsedEpisodeInfo),
+                Release = decision.Item.Release,
+                Title = decision.Item.Release.Title,
                 Added = DateTime.UtcNow
             });
 
             _eventAggregator.PublishEvent(new PendingReleasesUpdatedEvent());
         }
 
-        private void Delete(PendingRelease pendingRelease)
+        private void Delete(IPendingRelease pendingRelease)
         {
-            _repository.Delete(pendingRelease);
+            _repository.Delete(pendingRelease.Dao);
             _eventAggregator.PublishEvent(new PendingReleasesUpdatedEvent());
         }
 
-        private static Func<PendingRelease, bool> MatchingReleasePredicate(ReleaseInfo release)
+        private static Func<IPendingRelease, bool> MatchingReleasePredicate(ReleaseInfo release)
         {
             return p => p.Title == release.Title &&
-                   p.Release.PublishDate == release.PublishDate &&
-                   p.Release.Indexer == release.Indexer;
+                   p.RemoteItem.Release.PublishDate == release.PublishDate &&
+                   p.RemoteItem.Release.Indexer == release.Indexer;
         }
 
-        private int GetDelay(RemoteEpisode remoteEpisode)
+        private int GetDelay(RemoteItem remoteEpisode)
         {
-            var delayProfile = _delayProfileService.AllForTags(remoteEpisode.Series.Tags).OrderBy(d => d.Order).First();
+            var delayProfile = _delayProfileService.AllForTags(remoteEpisode.Media.Tags).OrderBy(d => d.Order).First();
             var delay = delayProfile.GetProtocolDelay(remoteEpisode.Release.DownloadProtocol);
             var minimumAge = _configService.MinimumAge;
 
             return new[] { delay, minimumAge }.Max();
         }
 
-        private void RemoveGrabbed(RemoteEpisode remoteEpisode)
+
+
+        private void RemoveGrabbedEpisode(RemoteEpisode remoteItem)
         {
             var pendingReleases = GetPendingReleases();
-            var episodeIds = remoteEpisode.Episodes.Select(e => e.Id);
+            var episodeIds = remoteItem.Episodes.Select(e => e.Id);
 
-            var existingReports = pendingReleases.Where(r => r.RemoteEpisode.Episodes.Select(e => e.Id)
-                                                             .Intersect(episodeIds)
-                                                             .Any())
-                                                             .ToList();
+            var existingReports = pendingReleases.OfType<EpisodePendingRelease>()
+                .Where(r => r.RemoteEpisode.Episodes.Select(e => e.Id)
+                    .Intersect(episodeIds)
+                    .Any())
+                .ToList();
 
             if (existingReports.Empty())
             {
                 return;
             }
 
-            var profile = remoteEpisode.Series.Profile.Value;
+            var profile = remoteItem.Media.Profile.Value;
 
             foreach (var existingReport in existingReports)
             {
-                var compare = new QualityModelComparer(profile).Compare(remoteEpisode.ParsedEpisodeInfo.Quality,
-                                                                        existingReport.RemoteEpisode.ParsedEpisodeInfo.Quality);
+                var compare = new QualityModelComparer(profile).Compare(remoteItem.Info.Quality,
+                                                                        existingReport.RemoteEpisode.Info.Quality);
 
                 //Only remove lower/equal quality pending releases
                 //It is safer to retry these releases on the next round than remove it and try to re-add it (if its still in the feed)
@@ -293,6 +344,11 @@ namespace NzbDrone.Core.Download.Pending
             }
         }
 
+        private void RemoveGrabbedMovie(RemoteMovie remoteItem)
+        {
+
+        }
+
         private void RemoveRejected(List<DownloadDecision> rejected)
         {
             _logger.Debug("Removing failed releases from pending");
@@ -300,7 +356,7 @@ namespace NzbDrone.Core.Download.Pending
 
             foreach (var rejectedRelease in rejected)
             {
-                var matching = pending.Where(MatchingReleasePredicate(rejectedRelease.RemoteEpisode.Release));
+                var matching = pending.Where(MatchingReleasePredicate(rejectedRelease.Item.Release));
 
                 foreach (var pendingRelease in matching)
                 {
@@ -310,14 +366,9 @@ namespace NzbDrone.Core.Download.Pending
             }
         }
 
-        private PendingRelease FindPendingRelease(int queueId)
+        private IPendingRelease FindPendingRelease(int queueId)
         {
-            return GetPendingReleases().First(p => p.RemoteEpisode.Episodes.Any(e => queueId == GetQueueId(p, e)));
-        }
-
-        private int GetQueueId(PendingRelease pendingRelease, Episode episode)
-        {
-            return HashConverter.GetHashInt31(string.Format("pending-{0}-ep{1}", pendingRelease.Id, episode.Id));
+            return GetPendingReleases().First(p => p.HasQueueId(queueId));
         }
 
         private int PrioritizeDownloadProtocol(Series series, DownloadProtocol downloadProtocol)
@@ -337,15 +388,15 @@ namespace NzbDrone.Core.Download.Pending
             _repository.DeleteBySeriesId(message.Series.Id);
         }
 
-        public void Handle(EpisodeGrabbedEvent message)
+        public void Handle(RemoteItemGrabbedEvent message)
         {
-            RemoveGrabbed(message.Episode);
+            message.Item.Do(RemoveGrabbedEpisode, RemoveGrabbedMovie);
         }
 
-        public void Handle(MovieGrabbedEvent message)
-        {
-            //RemoveGrabbed(message.Movie);
-        }
+        //public void Handle(MovieGrabbedEvent message)
+        //{
+        //    //RemoveGrabbed(message.Movie);
+        //}
 
         public void Handle(RssSyncCompleteEvent message)
         {
